backend/requirements.txt
#pip install -r requirements.txt
mysqlclient == 2.1.0
SQLAlchemy == 1.4.32
uvicorn == 0.17.6
fastapi == 0.75.0
PyMySQL == 0.9.3
alembic == 1.7.7
python-multipart == 0.0.5
passlib == 1.7.4
python-jose == 3.3.0
pytest == 7.1.2
bcrypt == 3.2.2
python-dotenv == 0.20.0
pytz == 2022.2.1
aiofiles == 0.8.0
requests == 2.28.1
# fastapi-pagination